Add setsockopt with SO_REUSEADDR
When remserial is used in conjunction with a USB to serial adapter and udev\rules.d to start remserial when the converter is plugged in, after a hard reset (eg. unplug/plug) or soft reset (usbreset) remserial fails to start because of TIME-WAIT state, with message "Couldn't bind port XXXX, aborting: Address already in use".
I have added setsockopt with SO_REUSEADDR option to overcome this issue.